Use automatic variables with [output/cppreference-doc-en-cpp.qch].
Replace pushd/popd with cd and a cp with ln.

Make [output/reference_cssless] an order-only prequesite.

As mentioned in the Makefile, qhelpgenerator needs to have all of the
referenced files (both via ref= and as <file/> entries) must be
relative to the *.qch file.  But this means that every time this
target runs, [output/reference_cssless] has the timestamp updated,
which causes [output/qch-files.xml] to be updated again, which is
eventually needed by the original target.  Since timestamps aren't
checked for order-only deps, this works around the problem.

Not particularly happy with this solution.  A better one might be to
create yet another working directory, link in directories from
reference_cssless, and use it.